Use `u64` rather than `usize` as "default integer" in library template
`usize` was renamed from `uint` in order to convey that it's not the
"default integer type". Guide new users to use integers with specific
bit width instead of `usize`.

+++ tests/testsuite/test.rs
@@ -5085,7 +5085,7 @@ fn cargo_test_set_out_dir_env_var() {
         .file(
             "src/lib.rs",
             r#"
-                pub fn add(left: usize, right: usize) -> usize {
+                pub fn add(left: u64, right: u64) -> u64 {
                     left + right
                 }
             "#,
